Fixed precision on numbers, added a native print function and fixed other natives
Integers are now displayed as integers when printed, and decimals are now displayed as float.
Added runtime errors for accessing an index out of range of an array and for an index that isn't an integer.
the native input function read() now converts input to their correct typing: number or string.
print is now a native function that receives a single argument which is printed.