Fixed type checking for arithmetic ops, fixed bug with scanning and bug with parsing
multiline comment style on the print token in the keywords dictionary in scanner.py caused to ignore all of the tokens following it.
Fixed type checking for arithmetic operations.
When I transferred the parsing error class to error.py I forgot to change it from being a class field, fixed it now.